Refactor error aggregators to use base streaming job
This refactoring was first attempted in 156089a which introduced serialization errors.
This is a second attempt, free from this bug.
