+++ src/test/scala/com/mozilla/telemetry/streaming/ExperimentsErrorAggregatorTest.scala
@@ -31,9 +31,7 @@ class ExperimentsErrorAggregatorTest extends FlatSpec with Matchers with DataFra
         ++ contentCrashes
         ++ TestUtils.generateMainMessages(k)).map(_.toByteArray).seq
 
-    val df = ErrorAggregator.aggregate(spark.sqlContext.createDataset(messages).toDF, raiseOnError = true,
-      ExperimentsErrorAggregator.defaultDimensionsSchema, ExperimentsErrorAggregator.defaultMetricsSchema,
-      ExperimentsErrorAggregator.defaultCountHistogramErrorsSchema)
+    val df = ErrorAggregator.aggregate(spark.sqlContext.createDataset(messages).toDF, raiseOnError = true)
 
     // 1 for each experiment (there are 2), and one for a null experiment
     df.count() should be (3)
